refactor: single handling enums
