--- src/contained-list/contained-list-kind.enum.ts
@@ -1,4 +1,0 @@
-export enum ContainedListKind {
-	OnPage = "on-page",
-	Disclosed = "disclosed"
-}
