+++ src/contained-list/index.ts
@@ -1,5 +1,4 @@
 export { ContainedList } from "./contained-list.component";
 export { ContainedListItem } from "./contained-list-item.component";
-export { ContainedListSize } from "./contained-list-size.enum";
-export { ContainedListKind } from "./contained-list-kind.enum";
+export { ContainedListSize, ContainedListKind } from "./contained-list.enums";
 export * from "./contained-list.module";
